backend/requirements.txt
fastapi>=0.104.1
sqlmodel>=0.0.16
uvicorn[standard]>=0.24.0
psycopg[binary]>=3.2.13
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.2
bcrypt>=4.0.1
# Phase 3 Technology Stack
openai>=1.0.0
pydantic>=2.5.0
mcp>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
# Better Auth JWT handling
python-jose>=3.3.0
PyJWT>=2.8.0